+++ js/modules/draw/TGraphPolarPainter.mjs
@@ -1,10 +1,13 @@
-import { settings, create } from '../core.mjs';
+import { settings, gStyle, create, BIT, clTPaveText, kTitle } from '../core.mjs';
 import { scaleLinear, select as d3_select, pointer as d3_pointer } from '../d3.mjs';
 import { DrawOptions, buildSvgCurve, makeTranslate } from '../base/BasePainter.mjs';
 import { ObjectPainter, getElementMainPainter } from '../base/ObjectPainter.mjs';
+import { TPavePainter } from '../hist/TPavePainter.mjs';
 import { ensureTCanvas } from '../gpad/TCanvasPainter.mjs';
 import { TooltipHandler } from '../gpad/TFramePainter.mjs';
+import { assignContextMenu, kNoReorder } from '../gui/menu.mjs';
 
+const kNoTitle = BIT(17);
 
 /**
  * @summary Painter for TGraphPolargram objects.
@@ -20,14 +23,37 @@ class TGraphPolargramPainter extends ObjectPainter {
       super(dom, polargram);
       this.$polargram = true; // indicate that this is polargram
       this.zoom_rmin = this.zoom_rmax = 0;
+      this.t0 = 0;
+      this.mult = 1;
+   }
+
+   /** @summary Returns true if fixed coordinates are configured */
+   isNormalAngles() {
+      const polar = this.getObject();
+      return polar?.fRadian || polar?.fGrad || polar?.fDegree;
+   }
+
+   /** @summary Set angles range displayed by the polargram */
+   setAnglesRange(tmin, tmax, set_obj) {
+      if (tmin >= tmax)
+         tmax = tmin + 1;
+      if (set_obj) {
+         const polar = this.getObject();
+         polar.fRwtmin = tmin;
+         polar.fRwtmax = tmax;
+      }
+      this.t0 = tmin;
+      this.mult = 2*Math.PI/(tmax - tmin);
    }
 
    /** @summary Translate coordinates */
-   translate(angle, radius, keep_float) {
+   translate(input_angle, radius, keep_float) {
+      // recalculate angle
+      const angle = (input_angle - this.t0) * this.mult;
       let rx = this.r(radius),
           ry = rx/this.szx*this.szy,
-          grx = rx * Math.cos(-angle - this.angle),
-          gry = ry * Math.sin(-angle - this.angle);
+          grx = rx * Math.cos(-angle),
+          gry = ry * Math.sin(-angle);
 
       if (!keep_float) {
          grx = Math.round(grx);
@@ -49,8 +75,10 @@ class TGraphPolargramPainter extends ObjectPainter {
    /** @summary Convert axis values to text */
    axisAsText(axis, value) {
       if (axis === 'r') {
-         if (value === Math.round(value)) return value.toString();
-         if (this.ndig>10) return value.toExponential(4);
+         if (value === Math.round(value))
+            return value.toString();
+         if (this.ndig > 10)
+            return value.toExponential(4);
          return value.toFixed(this.ndig+2);
       }
 
@@ -89,17 +117,16 @@ class TGraphPolargramPainter extends ObjectPainter {
 
    /** @summary Process mouse event */
    mouseEvent(kind, evnt) {
-      const layer = this.getLayerSvg('primitives_layer'),
-            interactive = layer.select('.interactive_ellipse');
-      if (interactive.empty()) return;
+//      const layer = this.getLayerSvg('primitives_layer'),
+//            interactive = layer.select('.interactive_ellipse');
+//      if (interactive.empty()) return;
 
       let pnt = null;
 
       if (kind !== 'leave') {
-         const pos = d3_pointer(evnt, interactive.node());
+         const pos = d3_pointer(evnt, this.draw_g.node());
          pnt = { x: pos[0], y: pos[1], touch: false };
       }
-
       this.processFrameTooltipEvent(pnt);
    }
 
@@ -124,7 +151,8 @@ class TGraphPolargramPainter extends ObjectPainter {
       // rmin -= delta*range;
       rmax += delta*range;
 
-      if ((rmin<polar.fRwrmin) || (rmax>polar.fRwrmax)) rmin = rmax = 0;
+      if ((rmin < polar.fRwrmin) || (rmax > polar.fRwrmax))
+         rmin = rmax = 0;
 
       if ((this.zoom_rmin !== rmin) || (this.zoom_rmax !== rmax)) {
          this.zoom_rmin = rmin;
@@ -155,7 +183,16 @@ class TGraphPolargramPainter extends ObjectPainter {
       }
 
       this.r = scaleLinear().domain([this.scale_rmin, this.scale_rmax]).range([0, this.szx]);
-      this.angle = polar.fAxisAngle || 0;
+
+      if (polar.fRadian) {
+         polar.fRwtmin = 0; polar.fRwtmax = 2*Math.PI;
+      } else if (polar.fDegree) {
+         polar.fRwtmin = 0; polar.fRwtmax = 360;
+      } else if (polar.fGrad) {
+         polar.fRwtmin = 0; polar.fRwtmax = 200;
+      }
+
+      this.setAnglesRange(polar.fRwtmin, polar.fRwtmax);
 
       const ticks = this.r.ticks(5),
             fontsize = Math.round(polar.fPolarTextSize * this.szy * 2);
@@ -165,7 +202,8 @@ class TGraphPolargramPainter extends ObjectPainter {
          nmajor = 8;
 
       this.createAttLine({ attr: polar });
-      if (!this.gridatt) this.gridatt = this.createAttLine({ color: polar.fLineColor, style: 2, width: 1, std: false });
+      if (!this.gridatt)
+         this.gridatt = this.createAttLine({ color: polar.fLineColor, style: 2, width: 1, std: false });
 
       const range = Math.abs(polar.fRwrmax - polar.fRwrmin);
       this.ndig = (range <= 0) ? -3 : Math.round(Math.log10(ticks.length / range));
@@ -183,6 +221,7 @@ class TGraphPolargramPainter extends ObjectPainter {
       }
 
       let exclude_last = false;
+      const pointer_events = this.isBatchMode() ? null : 'visibleFill';
 
       if ((ticks[ticks.length-1] < polar.fRwrmax) && (this.zoom_rmin === this.zoom_rmax)) {
          ticks.push(polar.fRwrmax);
@@ -191,13 +230,15 @@ class TGraphPolargramPainter extends ObjectPainter {
 
       return this.startTextDrawingAsync(polar.fRadialLabelFont, Math.round(polar.fRadialTextSize * this.szy * 2)).then(() => {
          for (let n = 0; n < ticks.length; ++n) {
-            let rx = this.r(ticks[n]), ry = rx/this.szx*this.szy;
+            let rx = this.r(ticks[n]),
+                ry = rx / this.szx * this.szy;
             this.draw_g.append('ellipse')
                .attr('cx', 0)
                .attr('cy', 0)
                .attr('rx', Math.round(rx))
                .attr('ry', Math.round(ry))
                .style('fill', 'none')
+               .style('pointer-events', pointer_events)
                .call(this.lineatt.func);
 
             if ((n < ticks.length-1) || !exclude_last) {
@@ -205,23 +246,24 @@ class TGraphPolargramPainter extends ObjectPainter {
                               text: this.format(ticks[n]), color: this.getColor(polar.fRadialLabelColor), latex: 0 });
             }
 
-            if ((nminor>1) && ((n < ticks.length-1) || !exclude_last)) {
+            if ((nminor > 1) && ((n < ticks.length - 1) || !exclude_last)) {
                const dr = (ticks[1] - ticks[0]) / nminor;
                for (let nn = 1; nn < nminor; ++nn) {
                   const gridr = ticks[n] + dr*nn;
                   if (gridr > this.scale_rmax) break;
-                  rx = this.r(gridr); ry = rx/this.szx*this.szy;
+                  rx = this.r(gridr);
+                  ry = rx / this.szx * this.szy;
                   this.draw_g.append('ellipse')
                      .attr('cx', 0)
                      .attr('cy', 0)
                      .attr('rx', Math.round(rx))
                      .attr('ry', Math.round(ry))
                      .style('fill', 'none')
+                     .style('pointer-events', pointer_events)
                      .call(this.gridatt.func);
                }
             }
          }
-
 
          return this.finishTextDrawing();
       }).then(() => {
@@ -231,7 +273,7 @@ class TGraphPolargramPainter extends ObjectPainter {
          const aligns = [12, 11, 21, 31, 32, 33, 23, 13];
 
          for (let n = 0; n < nmajor; ++n) {
-            const angle = -n*2*Math.PI/nmajor - this.angle;
+            const angle = -n*2*Math.PI/nmajor;
             this.draw_g.append('svg:path')
                 .attr('d', `M0,0L${Math.round(this.szx*Math.cos(angle))},${Math.round(this.szy*Math.sin(angle))}`)
                 .call(this.lineatt.func);
@@ -252,41 +294,43 @@ class TGraphPolargramPainter extends ObjectPainter {
          if (nminor > 1) {
             for (let n = 0; n < nmajor*nminor; ++n) {
                if (n % nminor === 0) continue;
-               const angle = -n*2*Math.PI/nmajor/nminor - this.angle;
+               const angle = -n*2*Math.PI/nmajor/nminor;
                this.draw_g.append('svg:path')
                    .attr('d', `M0,0L${Math.round(this.szx*Math.cos(angle))},${Math.round(this.szy*Math.sin(angle))}`)
                    .call(this.gridatt.func);
             }
          }
 
-         if (this.isBatchMode()) return;
+         if (this.isBatchMode())
+            return;
 
          TooltipHandler.assign(this);
 
-         const layer = this.getLayerSvg('primitives_layer');
-         let interactive = layer.select('.interactive_ellipse');
+         assignContextMenu(this, kNoReorder);
 
-         if (interactive.empty()) {
-            interactive = layer.append('g')
-                               .classed('most_upper_primitives', true)
-                               .append('ellipse')
-                               .classed('interactive_ellipse', true)
-                               .attr('cx', 0)
-                               .attr('cy', 0)
-                               .style('fill', 'none')
-                               .style('pointer-events', 'visibleFill')
-                               .on('mouseenter', evnt => this.mouseEvent('enter', evnt))
-                               .on('mousemove', evnt => this.mouseEvent('move', evnt))
-                               .on('mouseleave', evnt => this.mouseEvent('leave', evnt));
-         }
-
-         interactive.attr('rx', this.szx).attr('ry', this.szy);
-
-         d3_select(interactive.node().parentNode).attr('transform', this.draw_g.attr('transform'));
-
-         if (settings.Zooming && settings.ZoomWheel)
-            interactive.on('wheel', evnt => this.mouseWheel(evnt));
+         this.assignZoomHandler(this.draw_g);
       });
+   }
+
+   /** @summary Fill TGraphPolargram context menu */
+   fillContextMenuItems(menu) {
+      const pp = this.getObject();
+      menu.sub('Angle axis');
+      menu.addchk(pp.fRadian, 'Radian', flag => { pp.fRadian = flag; pp.fDegree = pp.fGrad = false; this.interactiveRedraw('pad', flag ? 'exec:SetToRadian()' : 'exec:SetTwoPi()'); }, 'Handle data angles as radian range 0..2*Pi');
+      menu.addchk(pp.fDegree, 'Degree', flag => { pp.fDegree = flag; pp.fRadian = pp.fGrad = false; this.interactiveRedraw('pad', flag ? 'exec:SetToDegree()' : 'exec:SetTwoPi()'); }, 'Handle data angles as degree range 0..360');
+      menu.addchk(pp.fGrad, 'Grad', flag => { pp.fGrad = flag; pp.fRadian = pp.fDegree = false; this.interactiveRedraw('pad', flag ? 'exec:SetToGrad()' : 'exec:SetTwoPi()'); }, 'Handle data angles as grad range 0..200');
+      menu.endsub();
+   }
+
+   /** @summary Assign zoom handler to element
+    * @private */
+   assignZoomHandler(elem) {
+      elem.on('mouseenter', evnt => this.mouseEvent('enter', evnt))
+          .on('mousemove', evnt => this.mouseEvent('move', evnt))
+          .on('mouseleave', evnt => this.mouseEvent('leave', evnt));
+
+      if (settings.Zooming && settings.ZoomWheel)
+         elem.on('wheel', evnt => this.mouseWheel(evnt));
    }
 
    /** @summary Draw TGraphPolargram */
@@ -316,16 +360,12 @@ class TGraphPolargramPainter extends ObjectPainter {
 
 class TGraphPolarPainter extends ObjectPainter {
 
-   /** @summary Redraw TGraphPolar */
-   redraw() {
-      this.drawGraphPolar();
-   }
-
    /** @summary Decode options for drawing TGraphPolar */
    decodeOptions(opt) {
       const d = new DrawOptions(opt || 'L');
 
-      if (!this.options) this.options = {};
+      if (!this.options)
+         this.options = {};
 
       Object.assign(this.options, {
           mark: d.check('P'),
@@ -335,39 +375,80 @@ class TGraphPolarPainter extends ObjectPainter {
           curve: d.check('C')
       });
 
+      if (d.check('A'))
+         this._draw_axis = true;
+
       this.storeDrawOpt(opt);
    }
 
+   /** @summary Update TGraphPolar with polargram */
+   updateObject(obj, opt) {
+      if (!this.matchObjectType(obj))
+         return false;
+
+      if (opt && (opt !== this.options.original))
+         this.decodeOptions(opt);
+
+      if (this._draw_axis && obj.fPolargram)
+         this.getMainPainter().updateObject(obj.fPolargram);
+
+      delete obj.fPolargram;
+      // copy all properties but not polargram
+      Object.assign(this.getObject(), obj);
+      return true;
+   }
+
+   /** @summary Redraw TGraphPolar */
+   redraw() {
+      return this.drawGraphPolar().then(() => this.updateTitle());
+   }
+
    /** @summary Drawing TGraphPolar */
-   drawGraphPolar() {
+   async drawGraphPolar() {
       const graph = this.getObject(),
             main = this.getMainPainter();
 
-      if (!graph || !main?.$polargram) return;
+      if (!graph || !main?.$polargram)
+         return;
 
-      if (this.options.mark) this.createAttMarker({ attr: graph });
-      if (this.options.err || this.options.line || this.options.curve) this.createAttLine({ attr: graph });
-      if (this.options.fill) this.createAttFill({ attr: graph });
+      if (this.options.mark)
+         this.createAttMarker({ attr: graph });
+      if (this.options.err || this.options.line || this.options.curve)
+         this.createAttLine({ attr: graph });
+      if (this.options.fill)
+         this.createAttFill({ attr: graph });
 
       this.createG();
+
+      if (this._draw_axis && !main.isNormalAngles()) {
+         const has_err = graph.fEX?.length;
+         let rwtmin = graph.fX[0],
+             rwtmax = graph.fX[0];
+         for (let n = 0; n < graph.fNpoints; ++n) {
+            rwtmin = Math.min(rwtmin, graph.fX[n] - (has_err ? graph.fEX[n] : 0));
+            rwtmax = Math.max(rwtmax, graph.fX[n] + (has_err ? graph.fEX[n] : 0));
+         }
+         rwtmax += (rwtmax - rwtmin) / graph.fNpoints;
+         main.setAnglesRange(rwtmin, rwtmax, true);
+      }
 
       this.draw_g.attr('transform', main.draw_g.attr('transform'));
 
       let mpath = '', epath = '';
-      const bins = [];
+      const bins = [], pointer_events = this.isBatchMode() ? null : 'visibleFill';
 
       for (let n = 0; n < graph.fNpoints; ++n) {
-         if (graph.fY[n] > main.scale_rmax) continue;
+         if (graph.fY[n] > main.scale_rmax)
+            continue;
 
          if (this.options.err) {
-            let pos1 = main.translate(graph.fX[n], graph.fY[n] - graph.fEY[n]),
-                pos2 = main.translate(graph.fX[n], graph.fY[n] + graph.fEY[n]);
-            epath += `M${pos1.grx},${pos1.gry}L${pos2.grx},${pos2.gry}`;
+            const p1 = main.translate(graph.fX[n], graph.fY[n] - graph.fEY[n]),
+                  p2 = main.translate(graph.fX[n], graph.fY[n] + graph.fEY[n]),
+                  p3 = main.translate(graph.fX[n] + graph.fEX[n], graph.fY[n]),
+                  p4 = main.translate(graph.fX[n] - graph.fEX[n], graph.fY[n]);
 
-            pos1 = main.translate(graph.fX[n] + graph.fEX[n], graph.fY[n]);
-            pos2 = main.translate(graph.fX[n] - graph.fEX[n], graph.fY[n]);
-
-            epath += `M${pos1.grx},${pos1.gry}A${pos2.rx},${pos2.ry},0,0,1,${pos2.grx},${pos2.gry}`;
+            epath += `M${p1.grx},${p1.gry}L${p2.grx},${p2.gry}` +
+                     `M${p3.grx},${p3.gry}A${p4.rx},${p4.ry},0,0,1,${p4.grx},${p4.gry}`;
          }
 
          const pos = main.translate(graph.fX[n], graph.fY[n]);
@@ -384,6 +465,7 @@ class TGraphPolarPainter extends ObjectPainter {
          if (this.options.fill) {
             this.draw_g.append('svg:path')
                 .attr('d', lpath + 'Z')
+                .style('pointer-events', pointer_events)
                 .call(this.fillatt.func);
          }
 
@@ -391,6 +473,7 @@ class TGraphPolarPainter extends ObjectPainter {
             this.draw_g.append('svg:path')
                 .attr('d', lpath)
                 .style('fill', 'none')
+                .style('pointer-events', pointer_events)
                 .call(this.lineatt.func);
          }
       }
@@ -399,6 +482,7 @@ class TGraphPolarPainter extends ObjectPainter {
          this.draw_g.append('svg:path')
                  .attr('d', buildSvgCurve(bins))
                  .style('fill', 'none')
+                 .style('pointer-events', pointer_events)
                  .call(this.lineatt.func);
       }
 
@@ -406,31 +490,39 @@ class TGraphPolarPainter extends ObjectPainter {
          this.draw_g.append('svg:path')
              .attr('d', epath)
              .style('fill', 'none')
+             .style('pointer-events', pointer_events)
              .call(this.lineatt.func);
       }
 
       if (mpath) {
          this.draw_g.append('svg:path')
                .attr('d', mpath)
+               .style('pointer-events', pointer_events)
                .call(this.markeratt.func);
+      }
+
+      if (!this.isBatchMode()) {
+         assignContextMenu(this, kNoReorder);
+         main.assignZoomHandler(this.draw_g);
       }
    }
 
    /** @summary Create polargram object */
-   createPolargram() {
-      const polargram = create('TGraphPolargram'),
-            gr = this.getObject();
+   createPolargram(gr) {
+      if (!gr.fPolargram)
+         gr.fPolargram = create('TGraphPolargram');
 
       let rmin = gr.fY[0] || 0, rmax = rmin;
+      const has_err = gr.fEY?.length;
       for (let n = 0; n < gr.fNpoints; ++n) {
-         rmin = Math.min(rmin, gr.fY[n] - gr.fEY[n]);
-         rmax = Math.max(rmax, gr.fY[n] + gr.fEY[n]);
+         rmin = Math.min(rmin, gr.fY[n] - (has_err ? gr.fEY[n] : 0));
+         rmax = Math.max(rmax, gr.fY[n] + (has_err ? gr.fEY[n] : 0));
       }
 
-      polargram.fRwrmin = rmin - (rmax-rmin)*0.1;
-      polargram.fRwrmax = rmax + (rmax-rmin)*0.1;
+      gr.fPolargram.fRwrmin = rmin - (rmax-rmin)*0.1;
+      gr.fPolargram.fRwrmax = rmax + (rmax-rmin)*0.1;
 
-      return polargram;
+      return gr.fPolargram;
    }
 
    /** @summary Provide tooltip at specified point */
@@ -443,19 +535,25 @@ class TGraphPolarPainter extends ObjectPainter {
 
       for (let n = 0; n < graph.fNpoints; ++n) {
          const pos = main.translate(graph.fX[n], graph.fY[n]),
-               dist2 = (pos.x-pnt.x)**2 + (pos.y-pnt.y)**2;
-         if (dist2 < best_dist2) { best_dist2 = dist2; bestindx = n; bestpos = pos; }
+               dist2 = (pos.grx - pnt.x)**2 + (pos.gry - pnt.y)**2;
+         if (dist2 < best_dist2) {
+            best_dist2 = dist2;
+            bestindx = n;
+            bestpos = pos;
+         }
       }
 
       let match_distance = 5;
-      if (this.markeratt?.used) match_distance = this.markeratt.getFullSize();
+      if (this.markeratt?.used)
+         match_distance = this.markeratt.getFullSize();
 
-      if (Math.sqrt(best_dist2) > match_distance) return null;
+      if (Math.sqrt(best_dist2) > match_distance)
+         return null;
 
       const res = {
          name: this.getObject().fName, title: this.getObject().fTitle,
-         x: bestpos.x, y: bestpos.y,
-         color1: this.markeratt?.used ? this.markeratt.color : this.lineatt.color,
+         x: bestpos.grx, y: bestpos.gry,
+         color1: (this.markeratt?.used ? this.markeratt.color : undefined) ?? (this.fillatt?.used ? this.fillatt.color : undefined) ?? this.lineatt?.color,
          exact: Math.sqrt(best_dist2) < 4,
          lines: [this.getObjectHint()],
          binindx: bestindx,
@@ -473,6 +571,59 @@ class TGraphPolarPainter extends ObjectPainter {
          res.lines.push(`error phi = ${main.axisAsText('phi', graph.fEX[bestindx])}`);
 
       return res;
+   }
+
+   /** @summary Only redraw histogram title
+     * @return {Promise} with painter */
+   async updateTitle() {
+      // case when histogram drawn over other histogram (same option)
+      if (!this._draw_axis)
+         return this;
+
+      const tpainter = this.getPadPainter()?.findPainterFor(null, kTitle, clTPaveText),
+            pt = tpainter?.getObject();
+
+      if (!tpainter || !pt)
+         return this;
+
+      const gr = this.getObject(),
+            draw_title = !gr.TestBit(kNoTitle) && (gStyle.fOptTitle > 0);
+
+      pt.Clear();
+      if (draw_title) pt.AddText(gr.fTitle);
+      return tpainter.redraw().then(() => this);
+   }
+
+
+   /** @summary Draw histogram title
+     * @return {Promise} with painter */
+   async drawTitle() {
+      // case when histogram drawn over other histogram (same option)
+      if (!this._draw_axis)
+         return this;
+
+      const gr = this.getObject(),
+            st = gStyle,
+            draw_title = !gr.TestBit(kNoTitle) && (st.fOptTitle > 0),
+            pp = this.getPadPainter();
+
+      let pt = pp.findInPrimitives(kTitle, clTPaveText);
+
+      if (pt) {
+         pt.Clear();
+         if (draw_title)
+            pt.AddText(gr.fTitle);
+         return this;
+      }
+
+      pt = create(clTPaveText);
+      Object.assign(pt, { fName: kTitle, fFillColor: st.fTitleColor, fFillStyle: st.fTitleStyle, fBorderSize: st.fTitleBorderSize,
+                           fTextFont: st.fTitleFont, fTextSize: st.fTitleFontSize, fTextColor: st.fTitleTextColor, fTextAlign: 22 });
+
+      if (draw_title)
+         pt.AddText(gr.fTitle);
+      return TPavePainter.draw(pp, pt, 'postitle')
+                         .then(p => { p?.setSecondaryId(this, kTitle); return this; });
    }
 
    /** @summary Show tooltip */
@@ -506,7 +657,8 @@ class TGraphPolarPainter extends ObjectPainter {
    /** @summary Process tooltip event */
    processTooltipEvent(pnt) {
       const hint = this.extractTooltip(pnt);
-      if (!pnt || !pnt.disabled) this.showTooltip(hint);
+      if (!pnt || !pnt.disabled)
+         this.showTooltip(hint);
       return hint;
    }
 
@@ -523,16 +675,16 @@ class TGraphPolarPainter extends ObjectPainter {
 
       let pr = Promise.resolve(null);
       if (!main) {
-         if (!graph.fPolargram)
-            graph.fPolargram = painter.createPolargram();
-         pr = TGraphPolargramPainter.draw(dom, graph.fPolargram);
+         // indicate that axis defined by this graph
+         painter._draw_axis = true;
+         pr = TGraphPolargramPainter.draw(dom, painter.createPolargram(graph));
       }
 
-      return pr.then(() => {
+      return pr.then(gram_painter => {
+         gram_painter?.setSecondaryId(painter, 'polargram');
          painter.addToPadPrimitives();
-         painter.drawGraphPolar();
-         return painter;
-      });
+         return painter.drawGraphPolar();
+      }).then(() => painter.drawTitle());
    }
 
 } // class TGraphPolarPainter